+++ DEFINES.py
@@ -1,27 +1,27 @@
 
 #python or matlab graph
-PYTHON_GRAPH = False
-GRAPH = True
+PYTHON_GRAPH = True
+GRAPH_MID = False #mid graphs
 
 #OPERETION MODS
 FLIP_MOD = False
-DELETE_MOD = False
+DELETE_MOD =False
 MIXED = False
-RANDOM_DELETIONS=True
+RANDOM_DELETIONS= True
 
 #count space as a mistake
 COUNT_SPACE_MISS = False
 
-NUMBER_OF_STRINGS_MIN = 2
-NUMBER_OF_STRINGS_MAX = None
-
-NUMBER_OF_DELETIONS_IN_STR_MIN = 0
-NUMBER_OF_DELETIONS_IN_STR_MAX= 5
-NUMBER_OF_FLIPS_IN_STR_MIN = 0
-NUMBER_OF_FLIPS_IN_STR_MAX = 5
-
-NUMBER_OF_TOTAL_MISTAKES_MAX = 5
-NUMBER_OF_TOTAL_MISTAKES_MIN = 0
 
 
+#PATHES
+MUSCLE_PATH ="/home/ubu/Yael/"#muscle3.8.31_i86linux64"   #Lab
+#MUSCLE_PATH ="C:/Users/moshab/Desktop/final project/muscle/"  #Boris_computer
+#MUSCLE_PATH ="C:/Users/boris7/Desktop/final project/muscle/"  #Boris_computer2
+#MUSCLE_PATH ="C:/Users/boris10/Desktop/projectCSE/muscle/"  #Boris_computer3
 
+FILES_PATH= "project/reasults/"#YAEL 18-10-18
+AVG_RES_FILE="avg_muscle.txt"#YAEL 18-10-18
+
+MUSCLE_IN_FILE = "in.txt"
+MUSCLE_OUT_FILE = "out.txt"
