--- PATHS.py
@@ -1,15 +1,0 @@
-
-
-#MUSCLE_PATH ="/home/ubu/Yael/"#muscle3.8.31_i86linux64"   #Lab
-#MUSCLE_PATH ="C:/Users/moshab/Desktop/final project/muscle/"  #Boris_computer
-MUSCLE_PATH ="C:/Users/boris7/Desktop/final project/muscle/"  #Boris_computer2
-#MUSCLE_PATH ="C:/Users/boris10/Desktop/projectCSE/muscle/"  #Boris_computer3
-
-
-
-FILES_PATH= "project/reasults/"#YAEL 18-10-18
-AVG_RES_FILE="avg_muscle.txt"#YAEL 18-10-18
-
-
-MUSCLE_IN_FILE = "in.txt"
-MUSCLE_OUT_FILE = "out.txt"
